requirements.txt
requests
beautifulsoup4
gspread
google-auth
lxml
googlesearch-python
